# Conflicts:
#	Proto/HotfixMessage.proto
#	Server/Hotfix/Module/Message/HotfixMessage.cs
#	Server/Hotfix/Module/Message/HotfixOpcode.cs
#	Unity/Assets/Res/Code/Hotfix.dll.bytes
#	Unity/Assets/Res/Code/Hotfix.mdb.bytes
#	Unity/Assets/Res/Code/Hotfix.pdb.bytes
#	Unity/Hotfix/Component/PlayerInfoComponent.cs
#	Unity/Hotfix/Module/Message/HotfixMessage.cs
#	Unity/Hotfix/Module/Message/HotfixOpcode.cs
#	Unity/Hotfix/Unity.Hotfix.csproj
#	Unity/Unity.Editor.Plugins.csproj
#	Unity/Unity.Editor.csproj
#	Unity/Unity.Plugins.csproj
#	Unity/Unity.csproj

+++ Server/Model/Njmj/Entity/Db/PlayerBaseInfo.cs
@@ -5,8 +5,8 @@ namespace ETModel
     {
         public long uid { get; set; }
         public string Name { get; set; }
-        public int GoldNum { get; set; }
-        public int WingNum { get; set; }
+        public long GoldNum { get; set; }
+        public long WingNum { get; set; }
         public string Icon { get; set; }
     }
 }
